pyproject.toml
[project]
name = "hive-agent-system"
version = "0.1.0"
description = "HIVE Agent System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "anthropic>=0.42.0",
    "pydantic>=2.10.0",
    "watchdog>=6.0.0",
    "pyyaml>=6.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[project.scripts]
hive = "packages.api.main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["packages*", "tests*"]
exclude = ["apps*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
